Refactor Frame.Data, Frame.SetData; add Frame.DataPtr.

+++ utils.go
@@ -107,15 +107,15 @@ func GenSyntVideoNewFrame(w, h int, fmt int32) chan *Frame {
 
 			for y := 0; y < h; y++ {
 				for x := 0; x < w; x++ {
-					frame.SetData(0, y*frame.LineSize(0)+x, x+y+i*3)
+					frame.SetData(0, y*frame.LineSize(0)+x, uint8(x+y+i*3))
 				}
 			}
 
 			// Cb and Cr
 			for y := 0; y < h/2; y++ {
 				for x := 0; x < w/2; x++ {
-					frame.SetData(1, y*frame.LineSize(1)+x, 128+y+i*2)
-					frame.SetData(2, y*frame.LineSize(2)+x, 64+x+i*5)
+					frame.SetData(1, y*frame.LineSize(1)+x, uint8(128+y+i*2))
+					frame.SetData(2, y*frame.LineSize(2)+x, uint8(64+x+i*5))
 				}
 			}
